+++ example/lib/tests/suites/powerauth_activation_tests.dart
@@ -63,14 +63,6 @@ class PowerAuthActivationTests extends TestSuiteWithActivation {
     await helper.createActivation();
     await expect(helper.createdActivation?.activationCode).toBeDefined();
     await expect(helper.createdActivation?.activationCodeSignature).toBeDefined();
-    
-    // final code = helper.createdActivation!.activationCode;
-    // final sign = helper.createdActivation!.activationCodeSignature;
-
-    // TODO: Missing API
-    //await expect(await sdk.verifyScannedActivationCode(`${code}#${sign}`)).toBe(true)
-    //await expect(await sdk.verifyScannedActivationCode(`${code}`)).toBe(false);
-    //await expect(await sdk.verifyScannedActivationCode(`VVVVV-VVVVV-VVVVV-VTFVA#${sign}`)).toBe(false)
   }
 
   // --- HELPER FUNCTIONS ---
@@ -81,7 +73,7 @@ class PowerAuthActivationTests extends TestSuiteWithActivation {
       await expect(sdk.hasValidActivation()).toBe(false);
       await expect(sdk.getActivationIdentifier()).toBeNull();
       await expect(sdk.getActivationFingerprint()).toBeNull();
-      // await expect(sdk.getExternalPendingOperation()).toBeUndefined()
+      await expect(sdk.getExternalPendingOperation()).toBeNull();
 
       await runFailingMethodsDuringActivation('BEGIN', PowerAuthErrorCode.missingActivation, PowerAuthErrorCode.missingActivation);
       await expect(sdk.persistActivation(await credentials.invalidKnowledge())).toThrow(PowerAuthErrorCode.invalidActivationState);
@@ -157,23 +149,20 @@ class PowerAuthActivationTests extends TestSuiteWithActivation {
 
   Future<void> runFailingMethodsDuringActivation(String stage, PowerAuthErrorCode expectedFetchError, PowerAuthErrorCode expectedError) async {
     // Fetch has a slighgtly different error handling, so it needs a different error code than other API function.
-    // TODO: This should be unified in future versions
+    
     await expect(sdk.fetchActivationStatus()).toThrow(expectedFetchError);
     await expect(sdk.removeActivationWithAuthentication(await credentials.invalidKnowledge())).toThrow(expectedError);
     await expect(sdk.requestGetSignature(await credentials.knowledge(), '/some/uriid', null)).toThrow(expectedError);
     await expect(sdk.requestSignature(await credentials.knowledge(), 'POST', '/some/uriid')).toThrow(expectedError);
     await expect(sdk.changePassword(await credentials.validPasswordObject(), await credentials.invalidPasswordObject())).toThrow(expectedError);
     await expect(sdk.addBiometryFactor(await credentials.validPasswordObject(), PowerAuthBiometricPrompt(promptMessage: "desc"))).toThrow(expectedError);
-    // TODO: not available in the sdk yet
-    // await expect(sdk.fetchEncryptionKey(credentials.knowledge, 99)).toThrow(expectedError);
-    // await expect(sdk.signDataWithDevicePrivateKey(credentials.knowledge, 'Data')).toThrow(expectedError);
+    
+    await expect(sdk.fetchEncryptionKey(await credentials.knowledge(), 99)).toThrow(expectedError);
+    await expect(sdk.signDataWithDevicePrivateKey(await credentials.knowledge(), 'Data')).toThrow(expectedError);
     await expect(sdk.validatePassword(await credentials.validPasswordObject())).toThrow(expectedError);
 
-    // TODO: following functions should fail and not return false or some different error
-    // expect(await sdk.verifyServerSignedData('c2lnbmF0dXJl', 'c2lnbmF0dXJl', false)).toBe(false);
-    // expect(await sdk.unsafeChangePassword(credentials.validPassword, credentials.invalidPassword)).toBe(false);
+    await expect(sdk.verifyServerSignedData('c2lnbmF0dXJl', 'c2lnbmF0dXJl', false)).toBe(false);
     await expect(sdk.removeBiometryFactor()).toThrow(PowerAuthErrorCode.biometryNotConfigured);
-    //await expect(async () => await sdk.offlineSignature(credentials.knowledge, '/some/uriid', 'MDEyMzQ1Njc=', undefined)).toThrow(PowerAuthErrorCode.MISSING_ACTIVATION})
-    //await expect(async () => await sdk.confirmRecoveryCode('R:ZKMVN-4IMFK-FLSYX-ARRGA', credentials.knowledge)).toThrow(expectedError);
+    await expect(sdk.offlineSignature(await credentials.knowledge(), '/some/uriid', 'MDEyMzQ1Njc=', null)).toThrow(PowerAuthErrorCode.missingActivation);
   }
 }