+++ example/lib/tests/suites/powerauth_biometrics_interactive_tests.dart
@@ -186,6 +186,6 @@ class PowerauthBiometricsInteractiveTests extends TestSuiteWithActivation {
   }
 
   Future<bool> _isFaceID() async {
-    return !Platform.isAndroid && (await sdk.getBiometryInfo()).biometryType == PowerAuthBiometryType.face;
+    return !Platform.isAndroid && (await PowerAuth.getBiometryInfo()).biometryType == PowerAuthBiometryType.face;
   }
 }